% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_class.R
\name{add_class}
\alias{add_class}
\title{Add a class to an object}
\usage{
add_class(x, new_class, location = "first")
}
\arguments{
\item{x}{the object to add a class to}

\item{new_class}{the new class or classes}

\item{location}{should the new class come before the old classes or after
them?}
}
\value{
the object, but with new classes
}
\description{
Adds a class to an object.
}
\examples{
a <- "Hello world!"
class(a)
b <- add_class(a, "awesome_text")
class(b)

# You can also add multiple classes to an object at once!
c <- add_class(a, c("class1", "class2"))
class(c)
}
