% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_tbl_as_adv_table.R
\name{as_adv_attr}
\alias{as_adv_attr}
\alias{as_adv_attr.adv_tbl}
\alias{as_adv_attr.data.frame}
\title{Coerce objet to an advanced attribute list}
\usage{
as_adv_attr(x, ...)

\method{as_adv_attr}{adv_tbl}(x, group_col = NULL)

\method{as_adv_attr}{data.frame}(x, nodeset_class, nodeset_name, node_name, group_col)
}
\arguments{
\item{x}{object to coerce to advanced edge list}

\item{...}{aditional arguments to pass on to methods}

\item{group_col}{column that denotes grouping. Currently only 1 group is supported.}

\item{nodeset_class}{column with the nodeset class}

\item{nodeset_name}{column with the nodeset name}

\item{node_name}{column with the node names}
}
\description{
Advanced attribute lists only contain information about the nodes and
include:
\itemize{
  \item nodeset_class
  \item nodeset_name
  \item node_name
  }
Additionally, there are three columns for every attribute - a column with the
name of the attribute, a column with the type that the attribute is, and a
column with the attribute value
}
