% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_tbl_as_adv_table.R
\name{as_adv_edge}
\alias{as_adv_edge}
\alias{as_adv_edge.adv_tbl}
\alias{as_adv_edge.data.frame}
\title{Coerce objet to an advanced edge list}
\usage{
as_adv_edge(x, ...)

\method{as_adv_edge}{adv_tbl}(x)

\method{as_adv_edge}{data.frame}(x, source_node_name, target_node_name)
}
\arguments{
\item{x}{object to coerce to advanced edge list}

\item{...}{aditional arguments to pass on to methods}

\item{source_node_name}{column with the source node name}

\item{target_node_name}{column with the target node name}
}
\description{
Advanced edge lists only contain information about the edges and include:
\itemize{
  \item source_nodeset_class
  \item source_nodeset_name
  \item sourcce_node_name
  \item target_nodeset_class
  \item target_nodeset_name
  \item target_node_name
  \item link_value
  \item network_name
}
}
