% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_tbl_as_adv_table.R
\name{as_adv_tbl}
\alias{as_adv_tbl}
\alias{as_adv_tbl.igraph}
\alias{as_adv_tbl.default}
\title{Coerce object to adv_tbl}
\usage{
as_adv_tbl(x, ...)

\method{as_adv_tbl}{igraph}(x)

\method{as_adv_tbl}{default}(x)
}
\arguments{
\item{x}{object to coerce to advanced table}

\item{...}{aditional arguments to pass on to methods}
}
\description{
Coerce object to adv_tbl
}
