% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot.R
\name{bar_plot}
\alias{bar_plot}
\title{Create Descriptive Statistics Bar Graph}
\usage{
bar_plot(x)
}
\arguments{
\item{x}{Data frame or object of the descript_stats class}
}
\value{
A ggplot graph of the descriptive statistics provided
by the input data frame
}
\description{
This function takes an input of a dataframe, specifically
the output of descript_stats. This data frame is then transformed
and filtered accordingly to make it graphable with ggplot2. The
ggplot graph is then returned.
}
