% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birthwt_clean}
\alias{birthwt_clean}
\title{Cleaned Birth Weight Dataset}
\format{
A data frame with 189 rows and 10 variables.
}
\source{
MASS package
}
\usage{
birthwt_clean
}
\description{
A cleaned version of the MASS::birthwt dataset, containing variables related
to maternal characteristics, health conditions, and infant birth weight.
}
\keyword{datasets}
