% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaned_protein_seq_aa.R
\name{cleaned_protein_seq_aa}
\alias{cleaned_protein_seq_aa}
\title{Clean protein sequences to keep only valid amino acids}
\usage{
cleaned_protein_seq_aa(sequences, warn = TRUE)
}
\arguments{
\item{sequences}{A character vector of protein sequences.}

\item{warn}{Logical, if \code{TRUE}, prints a warning for sequences with removed residues.}
}
\value{
A character vector of cleaned protein sequences with only valid uppercase amino acids.
}
\description{
Created because using \code{idpr} on real-world data sometimes produces errors
when sequences contain invalid residues. This function takes a character
vector of protein sequences and removes all characters that are not valid
one-letter amino acid codes. It converts sequences to uppercase.
}
\details{
Example of potential error avoided:
Error in sequenceCheck(sequence = sequence, method = "stop", outputType = "vector", :
Protein contains the following invalid residues

Valid amino acids are:
\code{A, R, N, D, C, E, Q, G, H, I, L, K, M, F, P, S, T, W, Y, V}
}
\examples{
sequences <- c("P12345" = "MKQTALV123$", "Q9XYZ1" = "ACD-ZEFG!")
cleaned <- cleaned_protein_seq_aa(sequences)
cleaned
}
