% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_stats.R
\name{compare_stats}
\alias{compare_stats}
\title{Comparing Statistics of Multiple DFs}
\usage{
compare_stats(x, y)
}
\arguments{
\item{x}{One data frame}

\item{y}{One data frame}
}
\value{
The ggplot graph of the comparative data
& prints data frame
}
\description{
This function takes in two data frames, specifically
those returned by descript_stats. This function then
compared the differences of the values, creates a new
graphable data frame with the methods, values, and
differences. This data frame is then used to create a
ggplot bar graph of the methods, values, and their differences.
This function both prints the graph and returns the data frame.
}
