% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descript_stats.R
\name{descript_stats}
\alias{descript_stats}
\title{Calculate Descriptive Statistics}
\usage{
descript_stats(x)
}
\arguments{
\item{x}{Numeric Vector}
}
\value{
An object with the class descript_stats, which prints a
data frame with the results
}
\description{
This function calculates descriptive statistics of mean,
median, mode, standard deviation, and variance. It then uses
then returns a structured data frame, coming from the descript_stats
object
}
\examples{
descript_stats(c(1,2,3,4,5))

}
