% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexel colors.R
\name{display_drexel_pal}
\alias{display_drexel_pal}
\title{Display a color palette}
\usage{
display_drexel_pal(palette, ...)
}
\arguments{
\item{palette}{Name of the color palette}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Creates a visual of the color palette of the drexel colors
}
