% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatech_colors.R
\name{display_gatech_pal}
\alias{display_gatech_pal}
\title{Display a color palette}
\usage{
display_gatech_pal(palette, ...)
}
\arguments{
\item{palette}{Name of the color palette}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Creates a visual of the color palette of the gatech colors
}
