% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexel colors.R
\name{drexel_cols}
\alias{drexel_cols}
\title{Function to extract drexel colors as hex codes}
\usage{
drexel_cols(...)
}
\arguments{
\item{...}{Character names of drexel_colors}
}
\description{
Function to extract drexel colors as hex codes
}
