% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drexel colors.R
\name{drexel_pal}
\alias{drexel_pal}
\title{Return function to interpolate a drexel color palette}
\usage{
drexel_pal(palette = "primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drexel_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a drexel color palette
}
