\name{Friedman-package}
\alias{Friedman-package}
\alias{Friedman}
\docType{package}
\title{
\packageTitle{Friedman}
}
\description{
\packageDescription{Friedman}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Friedman}
\packageIndices{Friedman}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Friedman}

Maintainer: \packageMaintainer{Friedman}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
