% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatech_colors.R
\docType{data}
\name{gatech_colors}
\alias{gatech_colors}
\title{Georgia Tech colors}
\format{
An object of class \code{character} of length 22.
}
\source{
<https://brand.gatech.edu/our-look/colors>
}
\usage{
gatech_colors
}
\description{
Color scheme from Georgia tech
}
\keyword{datasets}
