% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatech_colors.R
\name{gatech_cols}
\alias{gatech_cols}
\title{Function to extract gatech colors as hex codes}
\usage{
gatech_cols(...)
}
\arguments{
\item{...}{Character names of gatech_colors}
}
\description{
Function to extract gatech colors as hex codes
}
