% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatech_colors.R
\name{gatech_pal}
\alias{gatech_pal}
\title{Return function to interpolate a gatech color palette}
\usage{
gatech_pal(palette = "primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in gatech_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a gatech color palette
}
