% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_adv_data.R
\name{graph_from_adv_data}
\alias{graph_from_adv_data}
\title{Convert advanced format data into lists of edges and vertices}
\usage{
graph_from_adv_data(advanced_format_data, sourcenodesetname, sourcenodename,
  targetnodesetname, targetnodename, network, frequency, directed = TRUE)
}
\arguments{
\item{advanced_format_data}{data in advanced file format. Should be a
data.frame.it Should have the following columns, at the least: \itemize{
\item "Source nodeset name" \item "Source node name" \item "Target nodeset
name" \item "Target node name" \item "Frequency" \item "Network name" \item
"CHILD_ID" but if the complete advanced file is supplied, the other columns
will not be used.}}

\item{sourcenodesetname}{unqouted column name}

\item{sourcenodename}{unqouted column name}

\item{targetnodesetname}{unqouted column name}

\item{targetnodename}{unqouted column name}

\item{network}{unqouted column name}

\item{frequency}{unqouted column name}

\item{directed}{Logical scalar, whether or not to create a directed graph.
(passed to \code{igraph::graph_from_data_frame()})}
}
\value{
an igraph graph object
}
\description{
Converts advanced format network data (a la ORA) into an igraph object
}
\examples{
graph_from_adv_data(advanced_format_data)

}
\seealso{

}
\author{
Chris Friedman, \email{chris.s.friedman@gmail.com}
}
