% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_tbl_graph_from_adv_tbl.R
\name{graph_from_adv_tbl}
\alias{graph_from_adv_tbl}
\title{Convert adv_data in an igraph object}
\usage{
graph_from_adv_tbl(adv_data, directed = TRUE, ...)
}
\arguments{
\item{adv_data}{adv_tbl data created by \code{\link{adv_tbl()}}}
}
\value{
an igraph graph object
}
\description{
Converts advanced format network data (a la ORA) into an igraph object
}
\examples{
graph_from_adv_tbl(adv_data)

}
\author{
Chris Friedman, \email{chris.s.friedman@gmail.com}
}
