% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_all_items.R
\name{has_all_items}
\alias{has_all_items}
\title{Does a df grouping column have all possible instances?}
\usage{
has_all_items(df, column, col_items, cols_to_modify = NULL)
}
\arguments{
\item{df}{data frame to check}

\item{column}{column that should have all of the items specified in col_items}

\item{col_items}{items that should be present in col_items}

\item{cols_to_modify}{named list of columns that should be modified, where
list names correspond to column names and values correspond to new values
that should be in the added rows}
}
\value{
a df
}
\description{
check that a data frame has all of the specified items in the specified
column. if it does not have those items, add them. Also has functionality to
alter the added rows.
}
\examples{
df <- data.frame(groups = letters[1:5], value = rnorm(5))
has_all_items(df, groups, letters[1:10], list(value = 0))

}
