% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{Hemoglobin_subunit_beta}
\alias{Hemoglobin_subunit_beta}
\title{Hemoglobin Subunit Beta Sequence}
\format{
An \code{AAStringSet} object containing the full hemoglobin beta sequence.
}
\source{
\url{https://www.uniprot.org/uniprot/P68871}
}
\usage{
Hemoglobin_subunit_beta
}
\description{
Protein sequence for human Hemoglobin subunit beta (UniProt P68871) as an \code{AAStringSet}.
}
\examples{
library(Biostrings)
data(Hemoglobin_subunit_beta)
class(Hemoglobin_subunit_beta)
}
\keyword{datasets}
