% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{P53_Human_Protein}
\alias{P53_Human_Protein}
\title{P53 Human Protein Sequence}
\format{
An \code{AAStringSet} object containing the full P53 protein sequence.
}
\source{
\url{https://www.uniprot.org/uniprot/P04637}
}
\usage{
P53_Human_Protein
}
\description{
Protein sequence for human P53 (UniProt P04637) as an \code{AAStringSet}.
}
\examples{
library(Biostrings)
data(P53_Human_Protein)
class(P53_Human_Protein)
}
\keyword{datasets}
