% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friedman_plot.R
\name{plot.friedman_plot}
\alias{plot.friedman_plot}
\title{Plot method for a friedman_plot object}
\usage{
\method{plot}{friedman_plot}(x, ...)
}
\arguments{
\item{x}{a \code{friedman_plot} object}

\item{...}{ignored}
}
\value{
invisibly returns \code{x}
}
\description{
Re-renders the stored plot.
}
