% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bw_tbl.R
\name{print.bw_tbl}
\alias{print.bw_tbl}
\title{Custom print method for bw_tbl objects}
\usage{
\method{print}{bw_tbl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bw_tbl}.}

\item{...}{Additional arguments (ignored)}
}
\description{
This method provides a cleaner print output for the birth weight dataset.
}
