% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.descript_stats.R
\name{print.descript_stats}
\alias{print.descript_stats}
\title{descript_stats method}
\usage{
\method{print}{descript_stats}(x, ...)
}
\arguments{
\item{x}{Object of descript_stats class}

\item{...}{Additional arguments}
}
\value{
structured data frame of descript_stats output
}
\description{
This function creates a formatted version of descript_stats
by taking in the output, concatenating, and printing the
data frame.
}
