% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{quick_scatter}
\alias{quick_scatter}
\title{Scatter plot wrapper}
\usage{
quick_scatter(data, x, y)
}
\arguments{
\item{data}{a data frame}

\item{x}{x-axis column (unquoted)}

\item{y}{y-axis column (unquoted)}
}
\value{
a \code{friedman_plot} object
}
\description{
Scatter plot with a linear trend line using ggplot2.
}
\examples{
quick_scatter(mtcars, wt, mpg)
}
