% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Detect_macOS.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{bplapply}
\title{Objects exported from other packages}
\keyword{(helper)}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{BiocParallel}{\code{\link[BiocParallel]{bplapply}}}
}}

