% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{replace_column_w_mean}
\alias{replace_column_w_mean}
\title{replace column with mean}
\usage{
replace_column_w_mean(x)
}
\arguments{
\item{x}{data frame being cleaned/visualized}
}
\description{
replaces N/A values in a column with the column mean
}
