% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{Saccharomyces_cerevisiae_proteome}
\alias{Saccharomyces_cerevisiae_proteome}
\title{Saccharomyces cerevisiae Proteome}
\format{
An \code{AAStringSet} object containing all yeast protein sequences.
}
\source{
\url{https://www.uniprot.org/proteomes/UP000002311}
}
\usage{
Saccharomyces_cerevisiae_proteome
}
\description{
Full yeast proteome (UniProt UP000002311) as an \code{AAStringSet}.
}
\examples{
library(Biostrings)
data(Saccharomyces_cerevisiae_proteome)
class(Saccharomyces_cerevisiae_proteome)
}
\keyword{datasets}
