% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatech_colors.R
\name{scale_color_gatech}
\alias{scale_color_gatech}
\title{Color scale constructor for gatech colors}
\usage{
scale_color_gatech(palette = "primary", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in gatech_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for gatech colors
}
