% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_approx.R
\name{str_detect_approx}
\alias{str_detect_approx}
\alias{str_which_approx}
\alias{str_which_approx}
\alias{str_subset_approx}
\title{Approximate string matching, stringr style}
\usage{
str_detect_approx(string, pattern, ...)

str_which_approx(string, pattern, ...)

str_subset_approx(string, pattern, ...)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coearcible to one.}

\item{pattern}{Pattern to look for.}

\item{...}{see the documentation for \code{\link[base]{agrep()}} for other
arguments passed on to \code{agrep()} or \code{agrepl()}}
}
\value{
For \code{str_detect_approx()}, a logical vector. For
  \code{str_which_approx()} and \code{str_subset_approx()}, a character
  vector
}
\description{
Perform approximate string detection while using familiar stringr syntax.
  \code{str_detect_approx()} is a wrapper around base R \code{agrepl()}.
  \code{str_which_approx()} is a wrapper around base R \code{agrep()}.
  \code{str_subset_approx()} is a wrapper around
  \code{string[str_detect_approx(string, pattern)]}.
}
\examples{
friends <- c("Jon", "Ron", "Jack", "Mac", "jan")
str_detect_approx(friends, "Jon")
str_detect_approx(friends, "Jon", ignore.case = TRUE)
}
\seealso{
agrep()
}
