% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_idpr_results.R
\name{summarize_idpr_results}
\alias{summarize_idpr_results}
\title{Function to help summarize all protein computation results to include a protein-level summary and a Data set Level summary
writes it into Summary files
This function takes the output list from \code{idpr_parallel_processing()}
and produces a summarized table containing key metrics for a protein-level summary of each protein
and a data-set level summary, where the entire data set is taken into account.
Aggregates multiple IDPR computation outputs into summary tables for further analysis.}
\usage{
summarize_idpr_results(
  results,
  protein_outfile = "protein-level_summary.xlsx",
  dataset_outfile = "dataset-level_summary.xlsx",
  file_type = "xlsx"
)
}
\arguments{
\item{results}{A list of results from \code{idpr_parallel_processing()}.}

\item{protein_outfile}{Optional file path to save per-protein summary results as CSV.}

\item{dataset_outfile}{Optional file path to save dataset-level summary results as CSV.}

\item{file_type}{Character, format to write files, e.g., "csv".}
}
\value{
A list with two elements:
\describe{
\item{protein_summary}{Data frame with per-protein summary statistics.}
\item{dataset_summary}{Data frame with dataset-level summary statistics.}
}
}
\description{
Function to help summarize all protein computation results to include a protein-level summary and a Data set Level summary
writes it into Summary files
This function takes the output list from \code{idpr_parallel_processing()}
and produces a summarized table containing key metrics for a protein-level summary of each protein
and a data-set level summary, where the entire data set is taken into account.
Aggregates multiple IDPR computation outputs into summary tables for further analysis.
}
\examples{
\donttest{
sequences <- c(
  "P12345" = "MKTFFVAGA",
  "Q9XYZ1" = "ACDEFGHIKLMNPQRSTVWY"
)

idpr_functions <- c("idprofile", "chargeCalculationLocal")
valid_functions <- idpr_functions

if (requireNamespace("BiocParallel", quietly = TRUE)) {
  BPPARAM <- BiocParallel::SnowParam(1)
  results <- idpr_parallel_processing(
      idpr_functions = idpr_functions,
      valid_functions = valid_functions,
      input = sequences,
      BPPARAM = BPPARAM
  )
  str(results)
}
}
}
