% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_vector.R
\name{summarize_vector}
\alias{summarize_vector}
\title{Summarize a Numeric Vector}
\usage{
summarize_vector(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list containing the mean and median.
}
\description{
This function returns the mean and median of a numeric vector.
}
