% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friedman_plot.R
\name{summary.friedman_plot}
\alias{summary.friedman_plot}
\title{Summary of a friedman_plot object}
\usage{
\method{summary}{friedman_plot}(object, ...)
}
\arguments{
\item{object}{a \code{friedman_plot} object}

\item{...}{ignored}
}
\value{
invisibly returns \code{object}
}
\description{
Prints metadata about the plot and the variables used.
}
