% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{Tau_Human_Protein}
\alias{Tau_Human_Protein}
\title{Tau Human Protein Sequence}
\format{
An \code{AAStringSet} object containing the full Tau protein sequence.
}
\source{
\url{https://www.uniprot.org/uniprot/P10636}
}
\usage{
Tau_Human_Protein
}
\description{
Protein sequence for human Tau (UniProt P10636) as an \code{AAStringSet}.
}
\examples{
library(Biostrings)
data(Tau_Human_Protein)
class(Tau_Human_Protein)
}
\keyword{datasets}
