% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_tbl.R
\name{adv_tbl}
\alias{adv_tbl}
\alias{is_adv_tbl}
\title{Make Advanced Tables}
\usage{
adv_tbl(attributes, edges, directed = TRUE, group_col = NULL)
}
\arguments{
\item{attributes}{A data frame with the atributes ceated by \code{\link{as_adv_attr()}}}

\item{edges}{A data frame with the edges ceated by \code{\link{as_adv_edge()}}}

\item{directed}{Logical scalar - is the data directed (currently unused)}

\item{group_col}{column to group the data by}
}
\value{
an advanced table
}
\description{
Advanced Tables are a "long" way to look at network data where every node has
columns about their attributes. Additionally, all of the nodes and attributes
carry their own meta-data. This is a form that is made to make simple import
into ORA.
}
\examples{

}
