% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_idpr_results.R
\name{summarize_idpr_results}
\alias{summarize_idpr_results}
\title{Summarize IDPR Results}
\usage{
summarize_idpr_results(
  results,
  protein_outfile = "protein-level_summary.xlsx",
  dataset_outfile = "dataset-level_summary.xlsx",
  file_type = "xlsx"
)
}
\arguments{
\item{results}{List. Output from \code{idpr_parallel_processing()}.}

\item{protein_outfile}{Character. Path to save per-protein summary (optional).}

\item{dataset_outfile}{Character. Path to save dataset summary (optional).}

\item{file_type}{Character. File type to save ("csv" or "tsv"). Default is "csv".}
}
\value{
A list containing:
\describe{
\item{protein_summary}{Data frame with per-protein metrics.}
\item{dataset_summary}{Data frame summarizing the entire dataset.}
}
}
\description{
Takes the output list from \code{idpr_parallel_processing()} and produces summarized
tables containing key metrics for each protein and for the entire dataset.
}
\examples{
sequences <- c(
  "P12345" = "MKTFFVAGA",
  "Q9XYZ1" = "ACDEFGHIKLMNPQRSTVWY"
)
idpr_functions <- c("idprofile", "chargeCalculationLocal")
valid_functions <- idpr_functions
\dontrun{
  if (requireNamespace("BiocParallel", quietly = TRUE)) {
    BPPARAM <- BiocParallel::SnowParam(1)
    results <- idpr_parallel_processing(
      idpr_functions = idpr_functions,
      valid_functions = valid_functions,
      input = sequences,
      BPPARAM = BPPARAM
    )
    summarize_idpr_results(
      results,
      protein_outfile = "protein_summary.csv",
      dataset_outfile = "dataset_summary.csv"
    )
  }
}

}
